Resolve real path name for theme file
We are using alternative theme paths. Some of them are symlinks lik
gdm3.css. It points to a different directory and we need to ensure
assets are loaded from the real theme path then (assets path are
relative to it).
Expose thus realpath() to gjs and ensure we use the original file
itself when loading the stylesheet.