Split gnome-shell-extension-prefs out into its own binary package
It's not useful and arguably misleading to show this to everybody by
default, since it is only a tool for enabling/disabling extensions you
have already installed - and we don't provide any out of the box.

Perhaps it'd be useful for individual extensions to cause this to be
installed via Recommends.