Add layout-Try-to-allocate-before-getting-size-of-tracke.patch
To fix LP: #1919979 and also to re-fix LP: #1917939

Refixing LP: #1917939 in gnome-shell will allow us to remove the
mutter fix for that, which has caused a regression LP: #1922034 and
is being removed upstream for the same reason.

But this is only the first step and this commit only fixes
LP: #1919979 and LP: #1917939.