debian/gnome-shell-common.alternatives: Rename gdm3-theme to gdm-theme
gdm3 will be soon renamed to gdm, so given we're going to break the
users of gdm-theme.gresource this cycle now, we can safely do the rename
of the gdm gresource too.

Include a script that will remove all the previous alternatives and in
case a customization is done will warn about.