debian/patches: Support loading multiple Yaru theme variants
Yaru can provide multiple variants, we can support using more theme
variants in GNOME shell as it happens on Gtk apps.

So, make StSettings to compute the main theme and the chosen variant,
and we use those to pick the correct .css file.

We don't make difference between dark/light themes here, as we assume
that the shell theme will always be dark or light.